+++ packages/ssh/src/auth.test.ts
@@ -1,0 +1,22 @@
+import { assert, describe, it } from "@effect/vitest";
+import { Effect } from "effect";
+
+import { isSshAuthFailure } from "./auth.ts";
+
+describe("ssh auth", () => {
+  it.effect("detects ssh auth failures from common permission denied messages", () =>
+    Effect.sync(() => {
+      assert.equal(
+        isSshAuthFailure(
+          new Error(
+            "julius@100.65.180.100: Permission denied (publickey,password,keyboard-interactive).",
+          ),
+        ),
+        true,
+      );
+      assert.equal(isSshAuthFailure(new Error("Permission denied (publickey).")), true);
+      assert.equal(isSshAuthFailure(new Error("Connection timed out")), false);
+      assert.equal(isSshAuthFailure(new Error("mkdir: Permission denied")), false);
+    }),
+  );
+});
