+++ packages/ssh/src/config.test.ts
@@ -6,25 +6,16 @@ import { assert, describe, it } from "@effect/vitest";
 import { Effect } from "effect";
 
 import {
-  baseSshArgs,
   discoverSshHosts,
-  getLastNonEmptyOutputLine,
-  isSshAuthFailure,
   parseKnownHostsHostnames,
-  parseSshResolveOutput,
-  resolveRemoteT3CliPackageSpec,
   resolveSshConfigIncludePattern,
-} from "./ssh.ts";
-
-const tempDirectories: string[] = [];
+} from "./config.ts";
 
 function makeTempHomeDir(): string {
-  const directory = FS.mkdtempSync(Path.join(OS.tmpdir(), "t3-ssh-test-"));
-  tempDirectories.push(directory);
-  return directory;
+  return FS.mkdtempSync(Path.join(OS.tmpdir(), "t3-ssh-test-"));
 }
 
-describe("ssh", () => {
+describe("ssh config", () => {
   it.effect("discovers ssh config hosts across included files", () =>
     Effect.acquireUseRelease(
       Effect.sync(makeTempHomeDir),
@@ -132,102 +123,6 @@ describe("ssh", () => {
         resolveSshConfigIncludePattern("~/.ssh/config.d/*.conf", "/tmp/project", "/tmp/home"),
         "/tmp/home/.ssh/config.d/*.conf",
       );
-    }),
-  );
-
-  it.effect("parses resolved ssh config output into a target", () =>
-    Effect.sync(() => {
-      assert.deepEqual(
-        parseSshResolveOutput(
-          "devbox",
-          ["hostname devbox.example.com", "user julius", "port 2222", ""].join("\n"),
-        ),
-        {
-          alias: "devbox",
-          hostname: "devbox.example.com",
-          username: "julius",
-          port: 2222,
-        },
-      );
-    }),
-  );
-
-  it.effect("builds interactive ssh args without forcing batch mode", () =>
-    Effect.sync(() => {
-      assert.deepEqual(
-        baseSshArgs(
-          {
-            alias: "devbox",
-            hostname: "devbox.example.com",
-            username: "julius",
-            port: 2222,
-          },
-          { batchMode: "no" },
-        ),
-        ["-o", "BatchMode=no", "-o", "ConnectTimeout=10", "-p", "2222"],
-      );
-    }),
-  );
-
-  it.effect("resolves the remote t3 package spec from the desktop release channel", () =>
-    Effect.sync(() => {
-      assert.equal(
-        resolveRemoteT3CliPackageSpec({
-          appVersion: "0.0.17",
-          updateChannel: "latest",
-        }),
-        "t3@0.0.17",
-      );
-      assert.equal(
-        resolveRemoteT3CliPackageSpec({
-          appVersion: "0.0.17-nightly.20260415.44",
-          updateChannel: "nightly",
-        }),
-        "t3@0.0.17-nightly.20260415.44",
-      );
-      assert.equal(
-        resolveRemoteT3CliPackageSpec({
-          appVersion: "0.0.0-dev",
-          updateChannel: "nightly",
-          isDevelopment: true,
-        }),
-        "t3@nightly",
-      );
-      assert.equal(
-        resolveRemoteT3CliPackageSpec({
-          appVersion: "0.0.0-dev",
-          updateChannel: "latest",
-          isDevelopment: true,
-        }),
-        "t3@nightly",
-      );
-    }),
-  );
-
-  it.effect("reads the last non-empty ssh output line", () =>
-    Effect.sync(() => {
-      assert.equal(
-        getLastNonEmptyOutputLine(
-          ["Welcome to the host", "", '{"credential":"pairing-token"}', ""].join("\n"),
-        ),
-        '{"credential":"pairing-token"}',
-      );
-    }),
-  );
-
-  it.effect("detects ssh auth failures from common permission denied messages", () =>
-    Effect.sync(() => {
-      assert.equal(
-        isSshAuthFailure(
-          new Error(
-            "julius@100.65.180.100: Permission denied (publickey,password,keyboard-interactive).",
-          ),
-        ),
-        true,
-      );
-      assert.equal(isSshAuthFailure(new Error("Permission denied (publickey).")), true);
-      assert.equal(isSshAuthFailure(new Error("Connection timed out")), false);
-      assert.equal(isSshAuthFailure(new Error("mkdir: Permission denied")), false);
     }),
   );
 });
