+++ packages/ssh/src/config.ts
@@ -1,25 +1,14 @@
-import * as Crypto from "node:crypto";
 import * as FS from "node:fs";
 import * as OS from "node:os";
 import * as Path from "node:path";
 
-import type {
-  DesktopDiscoveredSshHost,
-  DesktopSshEnvironmentTarget,
-  DesktopUpdateChannel,
-} from "@t3tools/contracts";
+import type { DesktopDiscoveredSshHost } from "@t3tools/contracts";
 
-import { Data, Effect } from "effect";
+import { Effect } from "effect";
 
-export const DEFAULT_REMOTE_PORT = 3773;
+import { SshHostDiscoveryError } from "./errors.ts";
 
 const NO_HOSTS = [] as const;
-const PUBLISHABLE_T3_VERSION_PATTERN = /^\d+\.\d+\.\d+(?:-[0-9A-Za-z.-]+)?$/u;
-
-export class SshHostDiscoveryError extends Data.TaggedError("SshHostDiscoveryError")<{
-  readonly message: string;
-  readonly cause: unknown;
-}> {}
 
 function stripInlineComment(line: string): string {
   const hashIndex = line.indexOf("#");
@@ -183,106 +172,6 @@ function readKnownHostsHostnames(filePath: string): ReadonlyArray<string> {
   }
 
   return parseKnownHostsHostnames(FS.readFileSync(filePath, "utf8"));
-}
-
-export function parseSshResolveOutput(alias: string, stdout: string): DesktopSshEnvironmentTarget {
-  const values = new Map<string, string>();
-  for (const line of stdout.split(/\r?\n/u)) {
-    const trimmed = line.trim();
-    if (trimmed.length === 0) {
-      continue;
-    }
-    const [key, ...rest] = trimmed.split(/\s+/u);
-    if (!key || rest.length === 0 || values.has(key)) {
-      continue;
-    }
-    values.set(key, rest.join(" ").trim());
-  }
-
-  const hostname = values.get("hostname")?.trim() || alias;
-  const username = values.get("user")?.trim() || null;
-  const rawPort = values.get("port")?.trim() ?? "";
-  const parsedPort = Number.parseInt(rawPort, 10);
-
-  return {
-    alias,
-    hostname,
-    username,
-    port: Number.isInteger(parsedPort) ? parsedPort : null,
-  };
-}
-
-export function targetConnectionKey(target: DesktopSshEnvironmentTarget): string {
-  return `${target.alias}\u0000${target.hostname}\u0000${target.username ?? ""}\u0000${target.port ?? ""}`;
-}
-
-export function remoteStateKey(target: DesktopSshEnvironmentTarget): string {
-  return Crypto.createHash("sha256").update(targetConnectionKey(target)).digest("hex").slice(0, 16);
-}
-
-export function buildSshHostSpec(target: DesktopSshEnvironmentTarget): string {
-  const destination = target.alias.trim() || target.hostname.trim();
-  if (destination.length === 0) {
-    throw new Error("SSH target is missing its alias/hostname.");
-  }
-  return target.username ? `${target.username}@${destination}` : destination;
-}
-
-export function baseSshArgs(
-  target: DesktopSshEnvironmentTarget,
-  input?: { readonly batchMode?: "yes" | "no" },
-): string[] {
-  return [
-    "-o",
-    `BatchMode=${input?.batchMode ?? "no"}`,
-    "-o",
-    "ConnectTimeout=10",
-    ...(target.port !== null ? ["-p", String(target.port)] : []),
-  ];
-}
-
-export function normalizeSshErrorMessage(stderr: string, fallbackMessage: string): string {
-  const cleaned = stderr.trim();
-  return cleaned.length > 0 ? cleaned : fallbackMessage;
-}
-
-export function isSshAuthFailure(error: unknown): boolean {
-  const message = error instanceof Error ? error.message : String(error);
-  const normalized = message.toLowerCase();
-  return (
-    /permission denied \((?:publickey|password|keyboard-interactive|hostbased|gssapi-with-mic)[^)]*\)/u.test(
-      normalized,
-    ) ||
-    /authentication failed/u.test(normalized) ||
-    /too many authentication failures/u.test(normalized)
-  );
-}
-
-export function getLastNonEmptyOutputLine(stdout: string): string | null {
-  return (
-    stdout
-      .trim()
-      .split(/\r?\n/u)
-      .map((entry) => entry.trim())
-      .findLast((entry) => entry.length > 0) ?? null
-  );
-}
-
-export function resolveRemoteT3CliPackageSpec(input: {
-  readonly appVersion: string;
-  readonly updateChannel: DesktopUpdateChannel;
-  readonly isDevelopment?: boolean;
-}): string {
-  const appVersion = input.appVersion.trim();
-  if (!input.isDevelopment && PUBLISHABLE_T3_VERSION_PATTERN.test(appVersion)) {
-    return `t3@${appVersion}`;
-  }
-
-  if (input.isDevelopment) {
-    return "t3@nightly";
-  }
-
-  return input.updateChannel === "nightly" ? "t3@nightly" : "t3@latest";
 }
 
 export const discoverSshHosts = (input?: {
