+++ packages/ssh/src/errors.ts
@@ -1,0 +1,6 @@
+import { Data } from "effect";
+
+export class SshHostDiscoveryError extends Data.TaggedError("SshHostDiscoveryError")<{
+  readonly message: string;
+  readonly cause: unknown;
+}> {}
