+++ packages/ssh/src/index.ts
@@ -1,1 +1,5 @@
-export * from "./ssh.ts";
+export * from "./auth.ts";
+export * from "./command.ts";
+export * from "./config.ts";
+export * from "./errors.ts";
+export * from "./tunnel.ts";
