+++ packages/ssh/src/tunnel.ts
@@ -1,0 +1,6 @@
+export const DEFAULT_REMOTE_PORT = 3773;
+
+export function normalizeSshErrorMessage(stderr: string, fallbackMessage: string): string {
+  const cleaned = stderr.trim();
+  return cleaned.length > 0 ? cleaned : fallbackMessage;
+}
